requirements.txt
julius
numpy
torch==1.11.0
torchaudio==0.11.0
tqdm
dora-search
hydra-colorlog==1.1.0
hydra-core==1.1.0
diffq>=0.2.1